CIN-78 메인페이지 TOP 10 완료

+++ src/main/java/com/back/cinetalk/movie/dto/TopTenReviewDTO.java
@@ -17,7 +17,7 @@ public class TopTenReviewDTO {
 
     private String content;
 
-    private int likeCount;
+    private Long likeCount;
 
     private byte[] profile;
